feat: create Moderation module (Phase 3.3)\n\n- Move ModerationService from Services to Moderation namespace\n- Create ModerationModule with DI registration\n- Update Program.cs to use new module\n- Fix using directives in test files\n- All tests pass (1005/1010, 5 skipped due to API keys)\n\nProgress: 16/18 modules created, Program.cs reduced from 469 to ~320 lines
