+++ ClubDoorman.Test/Demo/TestKitDemoTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.TestInfrastructure;
 using NUnit.Framework;
