+++ ClubDoorman.Test/Integration/MessageHandlerBanBasicTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Models;
 using ClubDoorman.TestInfrastructure;
