+++ ClubDoorman.Test/Integration/MessageHandlerIntegrationTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using NUnit.Framework;
 using ClubDoorman.Handlers;
 using ClubDoorman.Services;
