+++ ClubDoorman.Test/Services/ModerationServiceSimpleTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Services;
 using ClubDoorman.Services.BanSystem;
 using ClubDoorman.TestInfrastructure;
