+++ ClubDoorman.Test/TestInfrastructure/StatisticsServiceTestFactory.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Services;
 using ClubDoorman.Services.BanSystem;
 using Microsoft.Extensions.Logging;
