+++ ClubDoorman.Test/TestKit/TestKit.UserJoinServiceBuilder.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Services.Messaging;
