+++ ClubDoorman.Test/Unit/Handlers/MessageHandlerAiCascadeAnalysisTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
