+++ ClubDoorman.Test/Unit/Handlers/MessageHandlerFakeTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Services;
 using ClubDoorman.Services.Core.Configuration;
