+++ ClubDoorman.Test/Unit/Handlers/MessageHandlerHandleAsyncBasicTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Test.TestInfrastructure;
 using ClubDoorman.Test.TestKit;
