+++ ClubDoorman.Test/Unit/Handlers/MessageHandlerHandleUserMessageTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Test.TestKit;
 using ClubDoorman.Models;
