+++ ClubDoorman.Test/Unit/Services/ChannelModerationServiceTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Infrastructure;
 using ClubDoorman.Services;
