+++ ClubDoorman.Test/Unit/Services/UserBanServiceTests.cs
@@ -1,3 +1,4 @@
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Models.Notifications;
 using ClubDoorman.Services;
 using ClubDoorman.Services.BanSystem;
