+++ ClubDoorman/Program.cs
@@ -8,6 +8,7 @@ using ClubDoorman.Services.Dispatcher;
 using ClubDoorman.Services.UserJoin;
 using ClubDoorman.Services.Notification;
 using ClubDoorman.Services.UserBan;
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Handlers;
 using ClubDoorman.Models.Logging;
 
@@ -104,6 +105,7 @@ public class Program
                 services.AddUserJoinServices();
                 services.AddNotificationServices();
                 services.AddUserBanServices();
+                services.AddModerationServices();
 
                 // Telegram Bot Client - создаем после регистрации IAppConfig
                 services.AddSingleton<TelegramBotClient>(provider =>
