+++ ClubDoorman/Services/Handlers/MessageHandler.cs
@@ -3,6 +3,7 @@ using System.Runtime.Caching;
 using System.Text;
 using Microsoft.Extensions.Logging;
 using ClubDoorman.Services.Commands;
+using ClubDoorman.Services.Moderation;
 using ClubDoorman.Infrastructure;
 using ClubDoorman.Models;
 using ClubDoorman.Models.Notifications;
