+++ ClubDoorman/Services/Moderation/ModerationModule.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace ClubDoorman.Services.Moderation;
+
+public static class ModerationModule
+{
+    public static IServiceCollection AddModerationServices(this IServiceCollection services)
+    {
+        // TODO: Добавить регистрацию сервисов если необходимо
+        return services;
+    }
+}
